markov/pyproject.toml
[tool.poetry]
name = "markovblog"
version = "0.1.0"
description = ""
authors = ["ssilvidi <steve.silvidi@rackspace.com>"]

[tool.poetry.dependencies]
python = "^3.9"
markovify = "^0.9.3"
spacy = "^3.1.3"
python-slugify = "^5.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
